convert program exit status returned by wait()
For some reason, the functions `wait` and `waitpidf` return an exit stauts
multiplied by 256 if it is greater than 0. For example, return 1024 if
the exit status of the program is 4, return 2560 if it is 10, and so on.
Therefore, the result needs to be divided by 256 before it is printed
out.